third_party/WebKit/LayoutTests/xr/xrView_match.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../vr/resources/fake-vr-displays.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/device/vr/vr_service.mojom.js"></script>
<script src="../vr/resources/mock-vr-service.js"></script>
<script src="../vr/resources/test-constants.js"></script>
<canvas id="webgl-canvas"></canvas>
<script src="../vr/resources/presentation-setup.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

xr_session_test( (t, session, mockService) => {
  // Session must have a baseLayer or frame requests will be ignored.
  let webglLayer = new XRWebGLLayer(session, gl);
  session.baseLayer = webglLayer;

  session.requestFrameOfReference("eyeLevel").then((frameOfRef) => {
    function onFrame(xrFrame) {
      t.step( () => {
        // Ensure that two views are provided.
        assert_not_equals(xrFrame.views, null);
        assert_equals(xrFrame.views.length, 2);

        let leftView = xrFrame.views[0];
        let rightView = xrFrame.views[1];

        // Ensure that the views are the right type.
        assert_true(leftView instanceof XRView);
        assert_true(rightView instanceof XRView);

        // Ensure that they have the expected eye enums.
        assert_equals(leftView.eye, "left");
        assert_equals(rightView.eye, "right");

        // Ensure they have the expected projection matrices.
        assert_not_equals(leftView.projectionMatrix, null);
        assert_not_equals(rightView.projectionMatrix, null);

        let displayLeftEye = fakeDisplays["Pixel"].leftEye;
        let displayRightEye = fakeDisplays["Pixel"].rightEye;

        let expectedLeftProjection = perspectiveFromFieldOfView(
            displayLeftEye.fieldOfView, session.depthNear, session.depthFar);
        let expectedRightProjection = perspectiveFromFieldOfView(
            displayRightEye.fieldOfView, session.depthNear, session.depthFar);

        assert_matrices_approx_equal(leftView.projectionMatrix, expectedLeftProjection);
        assert_matrices_approx_equal(rightView.projectionMatrix, expectedRightProjection);
      }, "XRView attributes match the expected values");
      t.done();
    }

    session.requestFrame(onFrame);
  });
}, fakeDisplays["Pixel"], { exclusive: true },
"XRPresentationFrame contains the expected views");

</script>
